% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.DESeq2.L2FC.R
\name{calc.DESeq2.L2FC}
\alias{calc.DESeq2.L2FC}
\title{Fold-change calculation for CRISPR screen}
\usage{
calc.DESeq2.L2FC(countsTable, metadata, meta.idnum = NULL,
  include.batch = F, plot.volcano = F, p.cutoff = 0.05,
  plot.MA = F, save = F)
}
\arguments{
\item{countsTable}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{metadata}{input dataframe containing sample names and other identifiers or data}

\item{meta.idnum}{vector containing the column numbers in metadata that represent (1) Cell line, (2) replicates, and (3) condition}

\item{include.batch}{logical to include replicates in the model}

\item{plot.volcano}{logical - do you want to show the volcano plot for each cell type}

\item{p.cutoff}{numeric - specified p-value cut-off}

\item{plot.MA}{logical - do you want to show the MA plot for each cell type}

\item{save}{logical - do you want to save the fold-change table to csv}
}
\value{
matrix containing Log2 fold-changes for each comparison
}
\description{
Calculates log2 fold-changes between samples of a specified variable using input of a count matrix and metadata dataframe
}
\examples{
L2FC <- calc.DESeq2.L2FC(countsTable, design.table, plot.MA = T, save = T)
...
}
