% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.seq.R
\name{calc.seq.depth}
\alias{calc.seq.depth}
\title{Sequencing depth}
\usage{
calc.seq.depth(
  countsTable,
  metadata,
  sample.id = "SampleID",
  print = T,
  save = F
)
}
\arguments{
\item{countsTable}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{metadata}{input dataframe containing sample names and other identifiers or data}

\item{sample.id}{string identifying column name in metadata containing sample IDs}

\item{print}{logical - do you want to print the  plot to plots}

\item{save}{logical - do you want to save the  plot to pdf}
}
\description{
bar plot visualizing sequencing depth
}
\examples{
y <- matrix(rnorm(100*9, mean = 10, sd = 1),100,9)
colnames(y) <- paste0('sample.',1:9)
metadata <- data.frame(sample = paste0('sample.',1:9), batch = c("A","A","A","B","B","B","C","C","C"))
calc.seq.depth(y)
...
}
\seealso{
\code{\link{ggplot}} which this function uses to plot
}
