% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.pca_function.R
\name{count.pca}
\alias{count.pca}
\title{PCA and Pearson correlation plots from counts table}
\usage{
count.pca(countsTable, metadata, identifier1 = NULL,
  identifier2 = NULL, identifier3 = NULL, batch = FALSE,
  batch.id = NULL, save = F)
}
\arguments{
\item{countsTable}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{metadata}{input dataframe containing sample names and other identifiers or data}

\item{identifier1}{string identifying column name of metadata with which to adjust color in PCA plot}

\item{identifier2}{string identifying column name of metadata with which to adjust alpha in PCA plot}

\item{identifier3}{string identifying column name of metadata with which to adjust shape in PCA plot}

\item{save}{logical - do you want to save the violin plot to pdf}
}
\value{
ggplot object of the violin plot
}
\description{
Basic comparison of CRISPR screen counts: PCA and Pearson correlation plots
}
\examples{
y <- matrix(rnorm(100*9, mean = 10, sd = 1),100,9)
y[,1:3] <- y[,1:3] + 5
metadata <- data.frame(sample = paste0('sample.',1:9), batch = c("A","A","A","B","B","B","C","C","C"))

count.pca(y, metadata, identifier1 = 'sample')
count.pca(y, metadata, identifier1 = 'sample', batch = T, batch.id = 'batch')
...
}
\seealso{
\code{\link{ggbiplot}} which this function uses to plot PCAs

\code{\link{ggplot}} which this function uses to plot Pearson correlations
}
