% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.violin_function.R
\name{count.violin}
\alias{count.violin}
\title{Count Distribution}
\usage{
count.violin(
  countsTable,
  metadata,
  identifier1 = "sample",
  identifier2 = NULL,
  transform = FALSE,
  cpm = FALSE,
  save = F
)
}
\arguments{
\item{countsTable}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{metadata}{input dataframe containing sample names and other identifiers or data}

\item{identifier1}{string identifying column name of metadata with which to adjust color in violin plot}

\item{identifier2}{string identifying column name of metadata with which to adjust alpha in violin plot}

\item{transform}{logical - do you want to log10 transform the counts}

\item{cpm}{logical - do you want to normalize the counts by sequencing depth}

\item{save}{logical - do you want to save the violin plot to pdf}
}
\value{
ggplot object of the violin plot
}
\description{
Violin plot visualizing read distribution
}
\examples{
y <- matrix(rnorm(100*9, mean = 10, sd = 1),100,9)
colnames(y) <- paste0('sample.',1:9)
metadata <- data.frame(sample = paste0('sample.',1:9), batch = c("A","A","A","B","B","B","C","C","C"),
time = rep(c(1,2,3),times = 3))
count.violin(y, metadata, identifier1 = 'sample', identifier2='time', transform = T)
...
}
\seealso{
\code{\link{ggplot}} which this function uses to plot
}
