% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential.genes.R
\docType{data}
\name{essential.genes}
\alias{essential.genes}
\title{Genes determined to be essential for cell fitness by Hart T et al. Cell (2015)}
\format{A character vector of gene names}
\source{
\url{https://www.cell.com/abstract/S0092-8674(15)01495-6}
}
\usage{
essential.genes
}
\description{
A dataset containing a vector of gene names for essential genes
}
\keyword{datasets}
