% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.genefromguide.R
\name{extract.genefromguide}
\alias{extract.genefromguide}
\title{Extraction of gene names}
\usage{
extract.genefromguide(data, neg.controls, split = "_")
}
\arguments{
\item{data}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{neg.controls}{input dataframe containing sample names and other identifiers or data}

\item{split}{string identifying cell line on which to filter counts}
}
\value{
new data with added column of gene names
}
\description{
Extraction of gene names from rownames that contain single guide RNA IDs
}
\examples{
new.L2FC <- extract.genefromguide(L2FC, neg.controls, split = '_')
}
