% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.enrich.R
\name{fgsea.enrich}
\alias{fgsea.enrich}
\title{FGSEA utility}
\usage{
fgsea.enrich(CSS, geneset, genome.db = Org.Hs.eg.db,
  gene.ids = "SYMBOL", p.value = 0.05, minSize = 15, maxSize = 500,
  nperm = 10000, top = 10, collapsepath = FALSE)
}
\arguments{
\item{CSS}{output CRISPR screen score matrix from create.CSS()}

\item{geneset}{input geneset}

\item{genome.db}{genome database from which to convert gene symbols to entrez ids}

\item{gene.ids}{type of gene ids in the geneset. Options are currently either 'symbol' or 'entrez'}

\item{p.value}{specified cutoff for pathway significance}

\item{minSize}{low cutoff for fgsea gene set size}

\item{maxSize}{high cutoff for fgsea gene set size}

\item{nperm}{number of permutations for fgsea analysis}

\item{top}{number of top pathways to display}

\item{collapsepath}{logical: collapse pathways using fgsea's collapsePathways function}
}
\value{
object of fgsea output
}
\description{
fgsea.enrich utilizes the package 'fgsea' to test for enrichment of user-given genesets in your gene list
}
\examples{
fgsea.enrich <- reactomepa.enrich(CSS.data[,2, drop=FALSE])
fgsea.enrich <- reactomepa.enrich(CSS.data[,c(2:4)])
...
}
\seealso{
\code{\link{fgsea}}
}
