% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.counts_function.R
\name{filterCounts}
\alias{filterCounts}
\title{Filtration of gRNAs with Low Counts}
\usage{
filterCounts(counts, thresh = 1, minsample.ids = 2)
}
\arguments{
\item{counts}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{thresh}{numerical cutoff for low counts}

\item{minsample.ids}{minimum number of samples for which counts must be present per gene}
}
\value{
new matrix of filtered counts
}
\description{
Filtration and visualization of counts
}
\examples{
y <- matrix(rnorm(100*9, mean = 10, sd = 1),100,9)
y[seq(1,100,10),] <- y[seq(1,100,10),]/20
counts.filtered <- filterCounts(y, thresh = 1, minsample.ids = 3)
...
}
