% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage.enrich.R
\name{gage.enrich}
\alias{gage.enrich}
\title{gage Geneset Enrichment Test}
\usage{
gage.enrich(CSS, geneset, genome.db = org.Hs.eg.db,
  gene.ids = "SYMBOL", species = "hsa", database = NULL)
}
\arguments{
\item{CSS}{output CRISPR screen score matrix from create.CSS()}

\item{geneset}{input geneset}

\item{genome.db}{genome database from which to convert gene symbols to entrez ids}

\item{gene.ids}{type of gene ids in the geneset. Options are currently either 'symbol' or 'entrez'}

\item{species}{species for pathview (if KEGG genesets are used)}

\item{database}{for visualizing pathways (use only if your database is KEGG)}
}
\value{
list of output from gage
}
\description{
gage.enrich utilizes the package 'gage' to test for enrichment of user-input gene sets
}
\examples{
enriched.kegg <- gage.enrich(results$CSS, kegg.genesets, gene.ids = 'entrez',
species = 'hsa', kegg = T)
enriched.msigdb <- gage.enrich(results$CSS, c2.canonical, gene.ids = 'symbol')
...
}
\seealso{
\code{\link{gage}, \link{pathview}}
}
