% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides2genes.R
\name{guides2genes}
\alias{guides2genes}
\title{CRISPR Screen Scoring}
\usage{
guides2genes(
  L2FC,
  neg.controls,
  essential.genes,
  save.plots = FALSE,
  print.plots = TRUE,
  save.tables = FALSE,
  split = "_"
)
}
\arguments{
\item{L2FC}{input matrix containing log2 fold-change data with gRNA ids as row names}

\item{neg.controls}{input vector containing gRNA ids for negative controls}

\item{essential.genes}{input vector containing gRNA ids for essential genes}

\item{save.plots}{save the plots to pdf (logical, default = FALSE)}

\item{print.plots}{print the plots (logical, default = TRUE)}

\item{save.tables}{save the CS and CSS tables (logical, default = FALSE)}

\item{split}{defines the regex used to split the guide RNA id to extract the gene name}
}
\value{
array containing additional object of a gene score table
}
\description{
Creates array containing log2 fold-change input, CRISPR scores, and CRISPR screen scores (similar to z-scores)
}
\examples{
test.array <- create.CSS(L2FC.filtered, neg.controls = neg.controls,
essential.genes = essential.genes)
...
}
\seealso{
\code{\link{L2FC.violin}, \link{ggplot2}, \link{ggbiplot}}
}
