% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.convert.R
\name{id.convert}
\alias{id.convert}
\title{ID converter}
\usage{
id.convert(CSS, genome.db = Org.Hs.eg.db, gene.ids = "SYMBOL")
}
\arguments{
\item{CSS}{output CRISPR screen score matrix from create.CSS()}

\item{genome.db}{genome database from which to convert gene symbols to other ids}

\item{gene.ids}{type of gene ids in the geneset}
}
\value{
object with converted rownames to the desired gene IDs
}
\description{
This function will take CSS with gene ids as rownames and convert them to a user-defined different ID using a user-given Annotation database
}
\examples{
data(CSS.data)
library(org.Hs.eg.db)
CSS.newids <- id.convert(CSS.data, genome.db = genome.db, gene.ids = gene.ids)
...
}
\seealso{
\code{\link[AnnotationDbi]{mapIds}}
}
