% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2FC.violin_function.R
\name{L2FC.violin}
\alias{L2FC.violin}
\title{Log2 Fold-Change Distribution}
\usage{
L2FC.violin(
  L2FC,
  input = "L2FC",
  variable = "time",
  save.plot = F,
  print.plot = T
)
}
\arguments{
\item{L2FC}{input matrix containing log2 fold-change values}

\item{input}{string identifying what type of input L2FC is (this function can also be used for visualizing CRISPR scores and CRISPR screen scores)}

\item{variable}{string identifying what variable was used to calculate L2FC values (for plot title purposes only)}

\item{save.plot}{logical - do you want to save the violin plot to pdf}

\item{print.plot}{logical - do you want to print the violin plot}
}
\value{
ggplot object of the violin plot
}
\description{
Violin plot visualizing log2 fold-change values
}
\examples{
L2FC.violin(L2FC.table)
}
\seealso{
\code{\link{ggplot}} which this function uses to plot
}
