% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.controls.R
\docType{data}
\name{neg.controls}
\alias{neg.controls}
\title{Names of negative controls for CRISPR screens using the Brunello library}
\format{A character vector of sgRNA names}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4744125/}
}
\usage{
neg.controls
}
\description{
A dataset containing a vector of names for negative control sgRNAs
}
\keyword{datasets}
