% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sigguides.R
\name{plot_sigguides}
\alias{plot_sigguides}
\title{Bidirectional barplot for sgRNA fold-change results}
\usage{
plot_sigguides(results, print = T, save = F, sig.cutoff = 0.05,
  cutoff = 1)
}
\arguments{
\item{results}{a calc.DESeq2.L2FC object}

\item{print}{logical - do you want to print the  plot to plots}

\item{save}{logical - do you want to save the  plot to pdf}

\item{sig.cutoff}{adjusted p-value cutoff for sgRNA fold-changes}

\item{fc.cutoff}{log2 fold-change cutoff for sgRNAs}
}
\description{
Visualize the number of significantly enriched/depleted sgRNAs
}
\examples{

...
}
\seealso{
\code{\link{ggplot}} which this function uses to plot
}
