% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{Upset plot for gene score results}
\usage{
plot_upset(genescore, save = FALSE, cutoff = 1, list.return = FALSE)
}
\arguments{
\item{genescore}{a guides2genes object}

\item{save}{logical - save the  plot to pdf}

\item{cutoff}{gene score cutoff for comparisons between contrasts (default = 1)}

\item{list.return}{logical - return a list containing gene names for every intersection}
}
\description{
Visualization of gene scores and overlaps
}
\examples{

...
}
\seealso{
\code{\link{UpSetR}, \link{Vennerable}} which this function uses to plot
}
