% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomepa.enrich.R
\name{reactomepa.enrich}
\alias{reactomepa.enrich}
\title{Reactome pathway analysis}
\usage{
reactomepa.enrich(
  CSS,
  genome.db = org.Hs.eg.db,
  pvalue = 0.05,
  combine = FALSE,
  combine.by = NULL,
  metric = "cutoff",
  cutoff = 1,
  save = FALSE,
  save.table = FALSE,
  showCategory = 15,
  font.size = 8,
  cex_label_category = 1,
  cex_label_gene = 1
)
}
\arguments{
\item{CSS}{output CRISPR screen score matrix from create.CSS()}

\item{genome.db}{genome database from which to convert gene symbols to entrez ids}

\item{pvalue}{specified cutoff for pathway significance (default = 0.05)}

\item{combine}{logical: do you want to combine the columns in the data? (default=FALSE)}

\item{combine.by}{vector separating the columns of your data into groups for comparing multiple groups}

\item{metric}{defines whether to look for enrichment based on a gene list of differentially expressed genes defined by a 'cutoff' (default) or by 'gsea' (one sample or fully combined samples only)}

\item{cutoff}{numeric cutoff to define differential expression (default = 1)}

\item{save}{logical: saves plots to pdf}

\item{save.table}{logical: saves tables of enriched pathways}

\item{showCategory}{number of categories to display}

\item{font.size}{font size of labels in dotplot and barplot}
}
\value{
list of (1) output from ReactomePA enrichment tests and (2) readable output
}
\description{
reactome.enrich utilizes the package 'ReactomePA' to test for enrichment of reactome pathways in your gene list
}
\examples{
reactome.enrich <- reactomepa.enrich(CSS.data[,1, drop=FALSE])
reactome.enrich <- reactomepa.enrich(CSS.data[,c(1:5)])
reactome.enrich <- reactomepa.enrich(CSS.data[,c(1:5)], combine = T)
reactome.enrich <- reactomepa.enrich(CSS.data[,1, drop = FALSE], metric = 'gsea')
...
}
\seealso{
\code{\link{ReactomePA}, \link{clusterProfiler}}
}
