% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct.R
\name{conduct}
\alias{conduct}
\title{Conduct: Single-function processing of CRISPR screen data}
\usage{
conduct(
  counts = NULL,
  metadata = NULL,
  sample.id = "SampleID",
  save.plot = T,
  save.table = T,
  identifier1 = NULL,
  identifier2 = NULL,
  identifier3 = NULL,
  batch = F,
  batch.id = NULL,
  transform = T,
  cpm = T,
  thresh = 1,
  minsample.ids = 2,
  meta.idnum = NULL,
  cutoff = 0.5,
  sig.cutoff = 0.05,
  verbose = T,
  top.labels = 30,
  neg.controls = NULL,
  essential.genes = NULL,
  split = "_",
  list.return = T,
  pathway.analysis = F,
  gene.db = NULL
)
}
\arguments{
\item{counts}{input matrix containing normalized gRNA counts with gRNA ids as row names}

\item{metadata}{input dataframe containing sample names and other identifiers or data including batch, treatments, etc.}

\item{sample.id}{string identifying column name in metadata containing sample IDs}

\item{save.plot}{logical - save the plot to pdf or print to screen (default saves everything to one pdf)}

\item{save.table}{logical - save the tables (default saves everything)}

\item{identifier1}{string identifying column name of metadata with which to adjust color in PCA plot}

\item{identifier2}{string identifying column name of metadata with which to adjust size in PCA plot}

\item{identifier3}{string identifying column name of metadata with which to adjust shape in PCA plot}

\item{batch}{logical - correct for batch effects (requires a batch.id input)}

\item{batch.id}{numerical - column of metadata that identifies the batch effect to remove}

\item{transform}{logical - do you want to log10 transform the counts}

\item{cpm}{logical - do you want to normalize the counts by sequencing depth}

\item{thresh}{numerical cutoff for low counts}

\item{minsample.ids}{minimum number of samples for which counts must be present per gene}

\item{meta.idnum}{vector containing the column numbers in metadata that represent (1) Cell line, (2) replicates, and (3) condition}

\item{cutoff}{gene score cutoff for comparisons between contrasts (default = 1)}

\item{sig.cutoff}{numeric - specified p-value cut-off}

\item{verbose}{TRUE/FALSE (default = TRUE)}

\item{top.labels}{numerical - number of top guides to label (by p-value)}

\item{neg.controls}{input vector containing gRNA ids for negative controls}

\item{essential.genes}{input vector containing gRNA ids for essential genes}

\item{split}{defines the regex used to split the guide RNA id to extract the gene name}

\item{list.return}{logical - return a list containing gene names for every intersection}

\item{pathway.analysis}{(TRUE/FALSE) continue with pathway analysis for all contrasts}

\item{gene.db}{genome database from which to convert gene symbols to other ids}
}
\description{
This function ties together all the basics of the COMPOSE package to give you a one-stop shop for processing data.
}
\details{
The required arguments to run the simplest analysis include counts, metadata, meta.idnum,
essential.genes, and neg.controls (if pathway.analysis = F).
}
\examples{

...
}
\seealso{
\code{\link{COMPOSE}} the COMPOSE package
}
